Web API big picture changes
The web API has been split into distributed services (which now have references to domain and message projects to make messages work properly).
RabbitMq has been used as the message provider and masstransit as the message library.
CoffeeSpace.OrderingApi.csproj has an implementation of the Saga pattern and StateMachine, which provides a way to manage order states.
In the near future it should also update the order state in the ordering db on each new order state. (Currently it doesn't).