MAUI Client Order Payment Implementation & Datadog Logging
Order.cs and OrderItem.cs in the Client folder now have observable properties. As a result, it created events that track when the aforementioned properties change.
Events were added to keep track of the order's status.
DbContext abstraction was removed since it was unnecessary.
Datadog logging has been enabled for all WEB APIs.